# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "sneat",
        "description": "Description de votre API",
        "contact": {
            "name": "Nom du contact",
            "email": "contact@api.com"
        },
        "license": {
            "name": "Licence de l'API",
            "url": "URL de la licence"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Connexion utilisateur",
                "description": "Retourne le donnee de user avec token authentification",
                "operationId": "login",
                "requestBody": {
                    "description": "Données du utilisatuer à envoyer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "JohnDoe@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "votre mot de passe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User connecter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "inscription de utilisateur",
                "description": "Inscription de nouvel utilisateur",
                "operationId": "register",
                "requestBody": {
                    "description": "Données du utilisatuer à envoyer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "JohnDoe@gmail.com"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "JohnDoe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "votre mot de passe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion avec success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
        "/api/friends/retirer/{user_id}": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Retirer une amis",
                "description": "Retirer une amis dans ma liste d'amis",
                "operationId": "retirer",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID du user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retirer une amis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/group/create": {
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "Creation d'une groupe",
                "description": "Creation d'une groupe ou je suis admin",
                "operationId": "creategroupe",
                "requestBody": {
                    "description": "Données de groupe a creer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "nom de groupe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Creation d'une groupe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/group/membres/{group_id}": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "Obtenir les membres d'un groupe",
                "description": "Retourne les membres d'un groupe en fonction de l'ID du groupe.",
                "operationId": "getMembreGroup",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "description": "ID du groupe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des membres du groupe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer votre_token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": [
                "read",
                "write"
            ]
        }
    ]
}